Move FetchClientMetadataDocument to pkg/oauthproto/cimd sub-package
Addresses jhrozek feedback on duplication:

- Create pkg/oauthproto/cimd sub-package which may import pkg/networking,
  eliminating three cases of duplicated logic:
  1. validateRedirectURI → now delegates to oauthproto.ValidateRedirectURI
     with RedirectURIPolicyStrict (RFC 8252 §8.4)
  2. isLoopbackHTTP → now uses oauthproto.IsLoopbackHost(parsed.Hostname())
     via URL.Parse, closing the localhost.evil.com prefix-match bypass
  3. cimdPrivateBlocks/isPrivateForCIMD → now uses networking.IsPrivateIP,
     eliminating the duplicated IP block list init

- pkg/oauthproto/cimd.go retains only ToolHiveClientMetadataDocumentURL
  and IsClientIDMetadataDocumentURL (Phase 1 code depends on these)

- HttpClientBuilder.WithPrivateIPs(false) is not used in newCIMDHTTPClient
  because it also blocks loopback (127.0.0.0/8), breaking the intentional
  http://localhost development exception. The custom DialContext is kept
  but uses networking.IsPrivateIP instead of the duplicated block list.
